+++ crates/stdio_server/src/session/handlers/on_move.rs
