+++ crates/stdio_server/src/session/mod.rs
@@ -1,6 +1,5 @@
 mod context;
-mod handler;
-pub mod impls;
+pub mod handlers;
 mod manager;
 mod providers;
 
@@ -10,7 +9,7 @@ use super::*;
 use crate::types::ProviderId;
 
 pub use self::context::SessionContext;
-pub use self::impls::on_move::{as_absolute_path, build_abs_path, OnMove, OnMoveHandler};
+pub use self::handlers::on_move::{as_absolute_path, build_abs_path, OnMove, OnMoveHandler};
 pub use self::manager::{Manager, NewSession, OpaqueSession};
 pub use self::providers::*;
